[Feature/BugFix] 	TaskEvent字段类型修改 [Solution Description]     	将TaskEvent taskPredictTime taskRealSpendTime 类型由float改为long [Other Info]     N/A
 modified:   app/src/main/java/com/ckt/ckttodo/database/EventTask.java

+++ app/src/main/java/com/ckt/ckttodo/database/EventTask.java
@@ -63,11 +63,11 @@ public class EventTask extends RealmObject {
     //任务开始时间
     private long taskStartTime;
     //任务预计花费时间
-    private float taskPredictTime;
+    private long taskPredictTime;
     //任务提醒时间
     private long taskRemindTime;
     //任务实际花费时间
-    private float taskRealSpendTime;
+    private long taskRealSpendTime;
     //任务所属计划
     private int planId;
 
@@ -143,11 +143,11 @@ public class EventTask extends RealmObject {
         this.taskStartTime = taskStartTime;
     }
 
-    public float getTaskPredictTime() {
+    public long getTaskPredictTime() {
         return taskPredictTime;
     }
 
-    public void setTaskPredictTime(float taskPredictTime) {
+    public void setTaskPredictTime(long taskPredictTime) {
         this.taskPredictTime = taskPredictTime;
     }
 
@@ -159,11 +159,11 @@ public class EventTask extends RealmObject {
         this.taskRemindTime = taskRemindTime;
     }
 
-    public float getTaskRealSpendTime() {
+    public long getTaskRealSpendTime() {
         return taskRealSpendTime;
     }
 
-    public void setTaskRealSpendTime(float taskRealSpendTime) {
+    public void setTaskRealSpendTime(long taskRealSpendTime) {
         this.taskRealSpendTime = taskRealSpendTime;
     }
